pyproject.toml
[project]
name = "google-ads-mcp"
version = "0.1.0"
description = "MCP Server for Google Ads API — full campaign management via Claude"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "google-ads>=25.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
